backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
google-generativeai==0.8.3
aiofiles==23.2.1
